Ajay-Patil/ImageSearch_Gen/requirements.txt
numpy
pillow
tensorflow
flask
openai==0.28
python-dotenv
tensorflow-cpu